+++ charts/the-web-cloud-platform/pipeline/run-implementation-tests.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env sh
 set -e
 
-helm test "the-web-cloud-platform-0"
+helm \
+  --namespace="default" \
+  test "the-web-cloud-platform-0"