+++ charts/the-web-cloud-platform/pipeline/run-system-tests.sh
@@ -1,10 +1,14 @@
 #!/usr/bin/env sh
 set -e
 
-testerPodName="$(kubectl get pod \
+testerPodName="$(kubectl \
+  --namespace="default" \
+  get pod \
   --output="jsonpath={.items[0].metadata.name}" \
   --selector="app.kubernetes.io/instance=the-web-cloud-platform-0,app.kubernetes.io/name=the-web-cloud-tester,app.kubernetes.io/component=tester" \
 )"
 
-kubectl exec "$testerPodName" -- \
+kubectl \
+  --namespace="default" \
+  exec "$testerPodName" -- \
   npm run target-api-tests